requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# CORS and middleware
fastapi-cors==0.0.6

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.3
numpy==1.25.2

# AI/ML libraries
openai==1.3.5
langchain==0.0.350
langchain-openai==0.0.2
transformers==4.36.0
torch==2.1.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Environment variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Validation and parsing
email-validator==2.1.0
python-multipart==0.0.6

# Date/time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
